\name{EcoHydrology-package}
\alias{EcoHydrology-package}
\alias{EcoHydrology}
\docType{package}
\title{
Principle equations used in Eco-Hydrology.
}
\description{
This package provides a flexible foundation for scientists, 
  engineers, and policy makers to base teaching exercises as well as for 
  more applied use to model complex eco-hydrological interactions. 
}
\details{
\tabular{ll}{
Package: \tab EcoHydrology\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2011-03-08\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
Included Energy Budget Functions:\cr
AtmosphericEmissivity\cr
EnvirEnergy\cr
EstCloudiness\cr
EvapHeat\cr
GroundHeat\cr
Longwave\cr
NetRad\cr
PotentialSolar\cr
RainHeat\cr
SatVaporDensity\cr
SensibleHeat\cr
Solar\cr
declination\cr
slopefactor\cr
solarangle\cr
solaraspect\cr
transmissivity\cr

Included Data Handeling Functions:\cr
get_gsod_stn\cr
build_gsod_forcing_data\cr

Included data:\cr
GSOD_history\cr

}
\author{
Fuka, DR, Walter, MT, and Easton, ZM\cr
Maintainer: Daniel Fuka <drf28@cornell.edu>
}
\references{
Anderson, E.A., 1968. Development and testing of snow pack energy balance equations. Water Resour. Res. 4, 19-37.
 
Anderson, E.A., 1976. A Point Energy and Mass Balance Model of a Snow Cover. U.S. Gov. Print. Off., Washington.

Ardia, David and Mullen, Katharine, 2009. DEoptim: Differential Evolution Optimization in R. R package version 2.0-3. 2010, 1. Available at: http://CRAN.R-project.org/package=DEoptim . Accessed 09/03, 2010.

Arnold, J.G., and N. Fohrer. 2005. SWAT2000: current capabilities and research opportunities in applied watershed modelling. Hydrological Processes 19:563-572.

Beven, K.J. 1997. TOPMODEL: A critique. Hydrological Processes 11:1069-1085.

Beven, K.J., 2001. Rainfall-Runoff Modelling : The Primer. J. Wiley, Chichester, UK; New York.

Bristow, K.L., and Campbell, G.S., 1984. On the relationship between incoming solar radiation and daily maximum and minimum temperature. Agric. For. Meteorol. 31, 159-166.

Campbell, G.S., Norman, J.M., 2000. An Introduction to Environmental Biophysics. Springer, New York, NY [u.a].

Corwin, D.L., Hopmans, J. and de Rooij, G.H.  From field-to landscape-scale vadose zone processes: scale issues, Vadose Zone J. 5 (2006), pp. 129–139.

Fuka, D.R., 1995. SPAW nitrogen fate routine : development and corroboration. Thesis (M.S.)--Washington State University, 1995.

Garnier, B., Ohmura, A., 1968. A Method of Calculating the Direct Shortwave Radiation Income of Slopes. J. Appl. Meteorol. 7, 796-800.

Hornik, K., 2010. The R FAQ.

Ihaka, R., Gentleman, R., 1996. R: A language for data analysis and graphics. Journal of computational and graphical statistics 5, 299-314.

Jensen, M., Burman, R., Allen, R., 1990. Evapotranspiration and irrigation water requirements.

Jensen, M.E., Burman, R.D., Allen, R.G., American Society of Civil Engineers. Committee on

Irrigation Water Requirements., 1990. Evapotranspiration and Irrigation Water Requirements : A Manual. The Society, New York, N.Y.

Monteith, J.L., Unsworth, M.H., 2008. Principles of Environmental Physics. Elsevier, Amsterdam [etc.].

Ndlovu, L.S., 1994. Weather data generation and its use in estimating evapotranspiration. Thesis (Ph. D.)--Washington State University, 1994.

Penman, H., 1948. Natural evaporation from open water, bare soil and grass. Proceedings of the Royal Society of London.Series A, Mathematical and Physical Sciences 193, 120-145.

Priestley, C., Taylor, R., 1972. On the assessment of surface heat flux and evaporation using large-scale parameters. Mon. Weather Rev. 100, 81-92.

Schneiderman, E.M. 2005. GWLF, In D. Radcliffe and M. Cabrera, eds. Modeling phosphorus in the environment. CRC Press.

Thornton, P.E., Hasenauer, H., White, M.A., 2000. Simultaneous estimation of daily solar radiation and humidity from observed temperature and precipitation: an application over complex terrain in Austria. Agric. For. Meteorol. 104, 255-271.

Walter, M.T., Brooks, E.S., McCool, D.K., King, L.G., Molnau, M., Boll, J., 2005. Process-based snowmelt modeling: does it require more input data than temperature-index modeling? Journal of Hydrology 300, 65-75.

Wigmosta, M.S., B. Nijssen, P. Storck, and D.P. Lettenmaier, 2002. The Distributed Hydrology Soil Vegetation Model, In Mathematical Models of Small Watershed Hydrology and Applications, V.P. Singh, D.K. Frevert, eds., Water Resource Publications, Littleton, CO., p. 7-42.

}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
}
