\name{GroundHeat}
\alias{GroundHeat}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Heat conducted to the bottom of a snowpack
}
\description{
The heat conducted to the bottom of a snowpack, assumed constant [kJ m-2 d-1]
}
\usage{
GroundHeat()
}
%- maybe also 'usage' for other objects documented here.
\author{
M. Todd Walter
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function(){
# the heat conducted to the bottom of a snowpack, assumed constant [kJ m-2 d-1]

return(173)
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
