\name{Longwave}
\alias{Longwave}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Daily Longwave Radiation
}
\description{
Daily longwave radiation based on the Sefan-Boltzman equation [kJ m-2 d-1]
}
\usage{
Longwave(emissivity, temp)
}
\arguments{
  \item{emissivity}{
emissivity [-]
}
  \item{temp}{
temperature of the emitting body [C]
}
}
\author{
M. Todd Walter
}

