\name{SatVaporPressure}
\alias{SatVaporPressure}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Saturated Vapor Pressure
}
\description{
Calculates the saturated vapor pressure (kPa) at a particular temperature (degrees Celcius)
}
\usage{
SatVaporPressure(T_C)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{T_C}{
%%     Temperature in degrees Celcius
}
}

\references{
Tetens (1930) : Tetens, V.O (1930) Uber einige meteorologische. Begriffe, Zeitschrift fur Geophysik. 6, 297-309
}
\author{
Josephine Archibald
}


