\name{swat_objective_function}
\alias{swat_objective_function}
\title{
A simple example objective function to be modified by user
}
\description{
A simple example objective function to be modified by user
}
\usage{
swat_objective_function(x, calib_range, calib_params,flowgage)
}
\arguments{
  \item{x}{
\code{x} A vector of adjustment factors to be fed into alter_params function
}
  \item{calib_range}{
\code{calib_range} This will in the future be a range for calibration.
}
  \item{calib_params}{
See other function to make sure you have this right.
}
  \item{flowgage}{
See other function to make sure you have this right.
}
}
\author{
Daniel Fuka
}
\examples{
\dontrun{
Sorry, this should be custom to your project. Look and determin what you want.
}
}
