\name{testSWAT2005}
\alias{testSWAT2005}
\title{
A function to test the numerical correctness of the SWAT2005 exe due to bad results on some CPUs.
}
\description{
A function to test the numerical correctness of the SWAT2005 exe due to bad results on some CPUs.
}
\usage{
testSWAT2005()
}
\author{
Daniel Fuka drf28@cornell.edu
}
