\name{testSWAT2012}
\alias{testSWAT2012}
\title{
A function to test the numerical correctness of the SWAT2012 exe due to bad results on some CPUs.
}
\description{
A function to test the numerical correctness of the SWAT2012 exe due to bad results on some CPUs.
}
\usage{
testSWAT2012()
}
\author{
Daniel Fuka drf28@cornell.edu
}
