\name{EcoHydRology-package}
\alias{EcoHydRology-package}
\alias{EcoHydRology}
\docType{package}
\title{
A community modeling foundation for Eco-Hydrology.
}
\description{
This package provides a flexible foundation for scientists, 
  engineers, and policy makers to base teaching exercises as well as for 
  more applied use to model complex eco-hydrological interactions. 
}
\details{
\tabular{ll}{
Package: \tab EcoHydRology\cr
Type: \tab Package\cr
Version: \tab 0.4.6\cr
Date: \tab 2011-12-22\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
Included Energy Budget Functions:\cr
AtmosphericEmissivity\cr
EnvirEnergy\cr
EstCloudiness\cr
EvapHeat\cr
GroundHeat\cr
Longwave\cr
NetRad\cr
PotentialSolar\cr
RainHeat\cr
SatVaporDensity\cr
SensibleHeat\cr
Solar\cr
declination\cr
slopefactor\cr
solarangle\cr
solaraspect\cr
transmissivity\cr

Included Data Handeling Functions:\cr
get_gsod_stn\cr
build_gsod_forcing_data\cr

Included data:\cr
GSOD_history\cr

}
\author{
Fuka DR, Walter MT, Archibald JA, Steenhuis, TS, and Easton ZM\cr
Maintainer: Daniel Fuka <drf28@cornell.edu>
}
\keyword{ package }
