\name{EstCloudiness}
\alias{EstCloudiness}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estimated Cloudiness 
}
\description{
Estimates the cloudiness of the atmosphere by scaling to atmospheric transmissivity
}
\usage{
EstCloudiness(Tx, Tn)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Tx}{
 maximum daily temperature [C]
}
  \item{Tn}{
 minimum daily temperature [C]
}
}
\author{
M. Todd Walter, Josephine Archibald
}
