\name{transmissivity}
\alias{transmissivity}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Transmissivity
}
\description{
Transmissivity
Fraction of direct solar radiation passing through the atmosphere based on the Bristow-Campbell eqn
}
\usage{
transmissivity(Tx, Tn)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
 \item{Tx}{
maximum daily temperature [C]
}
  \item{Tn}{
minimum daily temperature [C]
}
}
\details{
Can accept either single values or vectors for Tx and Tn.  
}
\references{
Bristow KL, Campbell GS. 1984. 
On the relationship between incoming solar radiation and daily maximum and minimum temperature. 
Agric Forest Meteorol; 31:150-166.
}
\author{
M. Todd Walter, Josephine Archibald
}
